Modify runmanager for DST analysis

+++ runmanager/module/singlerun.py
@@ -41,6 +41,14 @@ class SingleRun(object):
     self.__tag = tag
     self.__key = runinfo['key']
     self.__bin_path = runinfo['bin']
+    self.__is_dst = ('Dst' in self.__bin_path and
+                     'dstin' in runinfo and
+                     len(runinfo['dstin']) > 0)
+    if self.__is_dst:
+      self.__dstin_path = runinfo['dstin']
+    else:
+      self.__dstin_path = None
+    self.__dst_job = None
     self.__conf_path = runinfo['conf']
     self.__data_path = runinfo['data']
     self.__root_path = runinfo['root']
@@ -81,7 +89,8 @@ class SingleRun(object):
     self.__make_log()
     # self.__make_prefetch()
     self.__dump_init_info()
-    self.__make_element()
+    if self.__is_dst is not True:
+      self.__make_element()
     self.__ncomplete = 0
 
   #____________________________________________________________________________
@@ -155,27 +164,44 @@ class SingleRun(object):
     ''' Execute batch job. '''
     if self.__bjob_status is not None:
       return
-    self.__make_conf_list()
-    self.__make_root_list()
-    self.__make_log_list()
-    if self.__bjob_status is False:
-      return
-    for i, elem in enumerate(self.__elem_list):
-      if os.path.exists(self.__root_list[i]):
-        os.remove(self.__root_list[i])
-      job = bsub.BSub(self, elem,
-                      self.__conf_list[i],
-                      self.__root_list[i],
-                      self.__log_list[i])
-      job.execute()
-      self.__bjob_list.append(job)
-      self.__jobid_list.append(None)
-      self.__jobstat_list.append(job.get_status())
-    for i, item in enumerate(self.__bjob_list):
-      self.__dump_log(f'pid[bjob({i})]', item.get_process_id())
-    self.__dump_log(None, '_'*80)
-    self.__bjob_status = 0
-    # self.__status = 0
+    if self.__is_dst is not True: # User
+      self.__make_conf_list()
+      self.__make_root_list()
+      self.__make_log_list()
+      if self.__bjob_status is False:
+        return
+      for i, elem in enumerate(self.__elem_list):
+        if os.path.exists(self.__root_list[i]):
+          os.remove(self.__root_list[i])
+        job = bsub.BSub(self, elem,
+                        self.__conf_list[i],
+                        self.__root_list[i],
+                        self.__log_list[i])
+        job.execute()
+        self.__bjob_list.append(job)
+        self.__jobid_list.append(None)
+        self.__jobstat_list.append(job.get_status())
+      for i, item in enumerate(self.__bjob_list):
+        self.__dump_log(f'pid[bjob({i})]', item.get_process_id())
+      self.__dump_log(None, '_'*80)
+      self.__bjob_status = 0
+      # self.__status = 0
+    else: # Dst
+      command = shlex.split(f'bsub -q {self.__queue} ' +
+                            f'-o {self.__log_path} ')
+      command.extend([self.__bin_path,
+                      self.__conf_path])
+      command.extend(self.__dstin_path)
+      command.append(self.__root_path)
+      proc = subprocess.run(command,
+                            stdout=subprocess.PIPE,
+                            stderr=subprocess.PIPE,
+                            check=True)
+      job_id = bjob.BJob.read_job_id(proc.stdout.decode())
+      self.__dst_job = bjob.BJob(job_id)
+      self.__dump_log(f'jid[dst]', job_id)
+      self.__dump_log(None, '_'*80)
+      self.__bjob_status = 0
 
   #____________________________________________________________________________
   def finalize(self, keep_log=False):
@@ -279,6 +305,8 @@ class SingleRun(object):
   #____________________________________________________________________________
   def is_staged(self):
     ''' Is staged. '''
+    if self.__is_dst:
+      return True
     if self.__stage_status == 'STAGED':
       return True
     cmd = f'ghils {self.__data_path}'
@@ -401,10 +429,13 @@ class SingleRun(object):
       self.__status = 'UNKNOWN'
     if self.__bjob_status is None: # initial
       pass
-    elif self.__bjob_status == 0: # running
+    elif self.__bjob_status is 0: # running
       self.__status = 'BJOB-RUNNING'
     elif self.__bjob_status is True: # complete
-      self.__status = 'BJOB-DONE'
+      if self.__is_dst:
+        self.__status = 'DONE'
+      else:
+        self.__status = 'BJOB-DONE'
     elif self.__bjob_status is False: # failure
       self.__status = 'FAILED'
       return
@@ -596,6 +627,21 @@ class SingleRun(object):
     ''' Update job status. '''
     if (self.__bjob_status is not None and self.__bjob_status != 0):
       return
+    if self.__is_dst:
+      if self.__dst_job is None:
+        return
+      stat = self.__dst_job.get_status()
+      if stat == 'PEND' or stat == 'RUN':
+        pass
+      elif stat == 'DONE':
+        self.__bjob_status = True
+      elif stat == 'EXIT':
+        self.__bjob_status = False
+        self.__dump_log('error', 'merging error')
+        self.__dump_log(None, '_'*80)
+      else:
+        self.__merge_status = 'UNKNOWN'
+      return
     n_complete = 0
     for i, job in enumerate(self.__bjob_list):
       stat = job.get_status()
@@ -605,7 +651,7 @@ class SingleRun(object):
           self.__dump_log(f'time[bjob({i})]',
                          self.decode_time(job.get_run_time()))
           self.__dump_log(None, '_'*80)
-      elif stat is False and self.__jobstat_list[i] != stat:
+      elif stat is False and self.__jobstat_list[i] is 1:
         self.__bjob_status = False
         self.__dump_log('error', f'error at {job.get_tag()}')
         self.__dump_log(None, '_'*80)
@@ -660,9 +706,15 @@ class SingleRun(object):
     elif stat == 'STAGED':
       buff = 'staged'
     elif stat == 'BJOB-RUNNING':
-      buff = f'running({prog}/{nseg})'
+      if 'Dst' in data['bin']:
+        buff = 'running'
+      else:
+        buff = f'running({prog}/{nseg})'
     elif stat == 'BJOB-DONE':
-      buff = f'running({prog}/{nseg})'
+      if 'Dst' in data['bin']:
+        buff = 'running'
+      else:
+        buff = f'running({prog}/{nseg})'
     elif stat == 'MERGING':
       buff = f'merging({nseg})'
     elif stat == 'TERMINATED':
